backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.1
python-dotenv==1.0.0 